+++ client/src/scripts/apps/debug.ts
@@ -142,12 +142,18 @@ export class DebugTabApp extends TabApp {
         const stats = this.element!.querySelector("#debug-stats") as HTMLDivElement
         if (!stats) return
 
+        let prof=`
+main:${this.tab.game.clock.profiler.data[0]}<br>
+draw:${this.tab.game.clock.profiler.data[1]}<br>
+2:${this.tab.game.clock.profiler.data[2]}<br>
+        `
+
         stats.innerHTML = `
             FPS: ${Math.floor(1 / this.tab.game.delta_time)}<br>
             Ping: ${this.tab.game.client?.ping ?? 0}<br>
             X: ${this.tab.game.active_entity?.position.x}<br>
             Y: ${this.tab.game.active_entity?.position.y}<br>
-            Profilers ${Object.values(this.tab.game.clock.profiler.data)}
+            Profilers ${prof}
         `
     }