+++ common/engine/core/definition/definitions.ts
@@ -267,6 +267,11 @@ export interface AKeyFrameSpriteAction extends FrameTransform {
     image?: string
     fuser: string
 }
+export interface AKeyFrameTransformAction extends FrameTransform {
+    type: "transform"
+    fuser: string
+}
+
 
 export interface AKeyFrameTweenAction {
     type: "tween"
@@ -278,6 +283,7 @@ export interface AKeyFrameTweenAction {
 export type AKeyFrameAction =
     | AKeyFrameSpriteAction
     | AKeyFrameTweenAction
+    | AKeyFrameTransformAction
 export interface AKeyFrame{
     actions:AKeyFrameAction[]
     time:number